+++ src/lang-xpath/test/uk/co/reecedunn/intellij/plugin/intellij/tests/lexer/XPathSyntaxHighlighterTest.kt
@@ -354,6 +354,9 @@ class XPathSyntaxHighlighterTest {
         assertThat(highlighter.getTokenHighlights(XPathTokenType.K_ORELSE).size, `is`(1))
         assertThat(highlighter.getTokenHighlights(XPathTokenType.K_ORELSE)[0], `is`(XPathSyntaxHighlighterColors.KEYWORD))
 
+        assertThat(highlighter.getTokenHighlights(XPathTokenType.K_OTHERWISE).size, `is`(1))
+        assertThat(highlighter.getTokenHighlights(XPathTokenType.K_OTHERWISE)[0], `is`(XPathSyntaxHighlighterColors.KEYWORD))
+
         assertThat(highlighter.getTokenHighlights(XPathTokenType.K_PARAGRAPH).size, `is`(1))
         assertThat(highlighter.getTokenHighlights(XPathTokenType.K_PARAGRAPH)[0], `is`(XPathSyntaxHighlighterColors.KEYWORD))