+++ src/lang-xquery/test/uk/co/reecedunn/intellij/plugin/xquery/tests/psi/PluginConformanceTest.kt
@@ -1243,4 +1243,15 @@ private class PluginConformanceTest : ParserTestCase() {
         assertThat(conformance.conformanceElement, `is`(notNullValue()))
         assertThat(conformance.conformanceElement.elementType, `is`(XPathElementType.WILDCARD))
     }
+
+    @Test
+    @DisplayName("XQuery IntelliJ Plugin EBNF (114) OtherwiseExpr")
+    fun otherwiseExpr() {
+        val conformance = parse<PluginOtherwiseExpr>("a otherwise b")[0] as VersionConformance
+
+        assertThat(conformance.requiresConformance.size, `is`(1))
+        assertThat(conformance.requiresConformance[0], `is`(Saxon.VERSION_10_0))
+
+        assertThat(conformance.conformanceElement.elementType, `is`(XPathTokenType.K_OTHERWISE))
+    }
 }