+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Saxon 10 Syntax Extensions:
 
 1.  Support for wildcards in element and attribute-based kind tests.
 1.  Support for `.{ ... }` style context item inline function expressions.
+1.  Support for `otherwise` expressions.
 
 ## 2020